'fix' stupid compiler warnings
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@7291 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/mathed/formulabase.C
@@ -510,7 +510,7 @@ dispatch_result InsetFormulaBase::localDispatch(FuncRequest const & cmd)
 	}
 
 	case LFUN_PASTE: {
-		int n = 0;
+		size_t n = 0;
 		istringstream is(cmd.argument.c_str());
 		is >> n;
 		if (was_macro)
